Update patch set 1
Patch Set 1:

> This is a code change and requires a version bump. Yes you will
 > also need to change the type files to point to the latest versions.

 > This is a code change and requires a version bump. Yes you will
 > also need to change the type files to point to the latest versions.

Lusheng I'll need you to chime in here for direction.  I agree with Tommy that this is a code change and is a case for version to change.  I can bump the version and can modify the type file but you are the consumer of this so are you ready to change on your end?

Patch-set: 1
Reviewer: Gerrit User 1071 <1071@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0