Update patch set 2
Patch Set 2:

(1 comment)

My comment is that this codebase seems to share a large portion of code with the Docker plugin. This should probably be out into a common library that both import, not only because there are now two repos with lots of common code, but because it makes it hard to review what is actually in this PR; as I went through it, I recognized a lot of the code as being in the Docker plugin. So, I have no -1 or +1 here since it's hard to know what this plugin brings.

Patch-set: 2
Reviewer: Gerrit User 515 <515@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0