Don't change to https when use_tls is set
The k8s plugin's behavior of setting the health check probe type
scheme to https when it was http and the use_tls flag is set is
causing a problem for the Policy Handler.  This change leaves the
health check type alone, so, if it was http, it stays http even if
use_tls is set.

Change-Id: If004876f5142889abe0b24aeb4c0fdedaea3ab2b
Issue-ID: DCAEGEN2-591
Signed-off-by: Andrew Gauld <ag1282@att.com>