Update patch set 1
Patch Set 1: Code-Review-1

this change contradicts the approach that we use in ONAP for bringing the onap-dcae-dcaepolicy-lib.  

The agreement is that we do not change the API on the lib and docker/k8s plugins have the range for the version <3.0.0.  That allows us to have the lib hosted on public pypi and not to have the tsunami of plugin versions spreading through the blueprints in ONAP.  

Please abandon

Patch-set: 1
Reviewer: Gerrit User 1949 <1949@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1