Update patch set 1
Patch Set 1:

To me, the fact that the container eventually *does* come up tells me that this isn't a non-recoverable error. Non-recoverable to me is for scenarios where "it will never work in the current state" rather than  "it might work sometime in the future ". Why not make this a recoverable error and retry (I think we'd need a max_retries). 

Re whether to do automatic cleanups, this is consistent with how the CDAP broker used to function. If anything failed during an install, it would attempt to clean up "as much as it possibly could"

Patch-set: 1
Reviewer: Gerrit User 515 <515@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0