Update patch set 1
Patch Set 1:

The condition here is that the client has told the plugin to deploy a component and wait a certain amount of time for the component to become "ready" (i.e., to start passing its health check, aka readiness probe). The plugin has successfully made the deployment request to Kubernetes, has probed periodically to see if the component has become ready, and has reached the time limit without the component becoming ready.  In effect, we treat each probe that returns a "not ready" status as a recoverable error--we wait a bit, and try again.  Eventually we give up and throw a NonRecoverableError.

The plugin has a default maximum wait time.  It's possible to override this in the blueprint and even to set it to infinity.  

(Note that Kubernetes sends a success response to a request to deploy a set of entities very quickly, after doing very minimal checks.  It doesn't check for what seem to be very obvious problems, like asking to spin up containers from Docker images that don't exist.  It doesn't even pick a host where it will deploy things before it responds. Kubernetes is relentlessly optimistic.  The image for this container doesn't exist?  Keep trying.  Maybe someone will push it to the repo.  Maybe someone will come along and edit the container description and change the image name to one that will work.  It's a different way of doing things.)

Patch-set: 1
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0