Update patch set 3
Patch Set 3:

My reading of this code is that it creates a new node type (clamp.nodes.policy) derived from dcae.nodes.policy.  The new type adds a property to the node.  It also *replaces* the dcae.nodes.policy implementation of the "create" operation with a dummy operation that does nothing.  This means that putting a component in a blueprint and linking via a relationship to a clamp.nodes.policy node will *not* provide the component with any configuration information from policy.  Is this what's intended?

Patch-set: 3