Customize python import for kubernetes plugin
Cloudify manager fails to find the plugin without this helper.
The original code, cloudify-python-importer, is unmaintained.
It is just one module, licensed under Apache2.0 as well.
It also needs a python3 fix, hence the module code is added here.

Note that what the modification does should not be necessary,
and a proper root cause and fix needs to be found eventually.

Signed-off-by: Miroslav Los <miroslav.los@pantheon.tech>
Issue-ID: DCAEGEN2-1988
Change-Id: I28274dff902204362d7f5b6f97ac3381ff8b5411