Update patch set 3
Patch Set 3:

I've pulled down this code, built the plugin, loaded it onto a (Python 2.7-based) Cloudify Manager, and deployed a simple blueprint.   The deployment was successful--the earlier problem resulting from the missing import of "cloudify_importer" was resolved.

Suggested merging 99340 (bumping versions for all of the plugins), then patching this change to bump the version to 1.7.1, then merging this change.

Patch-set: 3
Label: Code-Review=0