Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

I have a couple nitpicky comments on handling of potentially unset/empty values, and how to fallback to defaults.
Perhaps another ticket would add code to ensure that the defaults are actually always in place, and so the rest of the code could be less defensive about the structure of the config.

Patch-set: 2
Label: Code-Review=-1