Update patch set 3
Patch Set 3:

As requested, I tested the nginx-svc-nodeport-ipv6.yaml blueprint in my environment and confirmed that 'concat' is being treated as a key, not an intrinsic function.  So it works.

I still believe it's a bad idea to use 'concat' instead of a meaningful name for the parameter--it's confusing. The nginx-svc-nodeport-ipv6.yaml blueprint is actually unrealistic, since it hardcodes the port numbers.  In a real blueprint, the port numbers are likely to come from inputs, and that will lead to lines that mix 'concat' as a key and 'concat' as an intrinsic function:

- concat: [{ concat: [{ get_input: [internalPort] },':', { get_input: [externalPort]} ]} ]

It seems to me that the following is much clearer:
- port_mapping: [{ concat: [{ get_input: [internalPort] },':', { get_input: [externalPort]} ]} ]

At the very least, if we going to keep this naming convention, we need really clear documentation explaining that concat is being used as a key, not an intrinsic function.  Anyone who is familiar with Cloudify blueprints but not familiar with the IPv6 feature in the k8splugin will be confused.

Patch-set: 3
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>