Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Change by itself looks fine, wondering though if we need version also to be bumped.. Possibly CM container rebuild should suffice to pull updated type file directly (under same plugin version)?
> 
> Jack - Any thoughts?

If we don't bump the version number, then when this change is merged, the updated types file will
replace the existing one in the Nexus raw repo.  The next time the CM container is rebuilt, it will
pull the updated types file. 

I think it's safest, though, to bump the version number here, then update /scripts/get-plugins.sh in the cm-container directory in the deployments repo (the script lists which versions of plugins/type files will get built into the cm-container so they can be uploaded to CM at deployment time), bump the version of the cm-container, release the new version of the cm-container, and--finally--update the OOM chart to use the new version.   It's a long process, but it guarantees that we will get the change we expect.

Patch-set: 1
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>